+++ src/proc/cuda_proc.h
@@ -4,8 +4,8 @@
 class CudaImageProc : public ImageProc {
 public:
   CudaImageProc();
-  virtual void Brighten(double value) override;
-  virtual void Sharpen(double value) override;
+  virtual void Brighten(float value) override;
+  virtual void Sharpen(float value) override;
   virtual bool IsSupported() const override;
   virtual std::string Name() const override { return "cuda"; }
   virtual ImageIOResult LoadImage(std::string filename) override;
